Mapping and validation features were added.

+++ list_api/Repository/ProductRepository.cs
@@ -1,6 +1,8 @@
 ﻿using list_api.Models;
 using list_api.Repository.Interface;
 using list_api.Data;
+using System.Net;
+using System.Web.Http;
 namespace list_api.Repository {
 	public class ProductRepository : IProductRepository {
 		private readonly IListApiDbContext context;
@@ -9,7 +11,7 @@ namespace list_api.Repository {
 		}
 		public Product Create(Product product) { // Creating a product.
 			Product? product_created = new Product() { Name = product.Name, Description = product.Description, IDCategory = product.IDCategory, Price = product.Price };
-			if (context.Products.Any(p => p.Name == product.Name)) throw new InvalidOperationException("Product already exists.");
+			if (context.Products.Any(p => p.Name == product.Name)) throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.Conflict) { ReasonPhrase = "Product already exists." });
 			context.Products.Add(product_created);
 			context.SaveChanges();
 			return product_created;
@@ -17,6 +19,7 @@ namespace list_api.Repository {
 		public Product? Delete(int id) { // Deleting a product.
 			Product? product_deleted = context.Products.FirstOrDefault(p => p.ID == id);
 			if (product_deleted != null) {
+				if (context.ListProducts.Any(lp => lp.IDProduct == id)) throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.Conflict) { ReasonPhrase = "Product exists in a list." });
 				context.Products.Remove(product_deleted);
 				context.SaveChanges();
 			}
@@ -34,7 +37,7 @@ namespace list_api.Repository {
 		public Product? Update(int id, Product product) { // Updating a product.
 			Product? product_updated = context.Products.FirstOrDefault(p => p.ID == id);
 			if (product_updated != null) {
-				if (context.Products.Any(p => p.Name == product.Name)) throw new InvalidOperationException("Product already exists.");
+				if (context.Products.Any(p => p.Name == product.Name)) throw new HttpResponseException(new HttpResponseMessage(HttpStatusCode.Conflict) { ReasonPhrase = "Product already exists." });
 				product_updated.Name = product.Name;
 				product_updated.Description = product.Description;
 				product_updated.IDCategory = product.IDCategory;
